Moved Extensions definitions to Util.
