+++ TR123/TR123.csproj
@@ -81,7 +81,6 @@
     <Compile Include="GameAddresses.cs" />
     <Compile Include="GameData.cs" />
     <Compile Include="GameMemory.cs" />
-    <Compile Include="ProcessExtensions.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Stats\GameStats.cs" />
     <Compile Include="Stats\LevelStats.cs" />
