+++ TR456/ComponentSettings.cs
@@ -4,6 +4,7 @@ using System.Drawing;
 using System.Linq;
 using System.Reflection;
 using System.Windows.Forms;
+using Util;
 
 namespace TR456;
