+++ TR456/TR456.csproj
@@ -85,7 +85,6 @@
     <Compile Include="Enums\Tr5LevelSection.cs" />
     <Compile Include="Enums\Tr6Level.cs" />
     <Compile Include="Enums\TransitionDirection.cs" />
-    <Compile Include="Extensions.cs" />
     <Compile Include="GameData.cs" />
     <Compile Include="GameMemory.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
