+++ TR456/VersionDetector.cs
@@ -2,6 +2,7 @@
 using System.Collections.Immutable;
 using System.Diagnostics;
 using System.Linq;
+using Util;
 
 namespace TR456;
