+++ TombRaiderIV/ComponentSettings.cs
@@ -5,6 +5,7 @@ using System.Drawing;
 using System.Linq;
 using System.Windows.Forms;
 using LaterClassicUtil;
+using Util;
 
 namespace TR4;
