+++ TombRaiderIV/TR4.csproj
@@ -60,7 +60,6 @@
       <SubType>UserControl</SubType>
     </Compile>
     <Compile Include="GameData.cs" />
-    <Compile Include="Enums\EnumExtensions.cs" />
     <Compile Include="Enums\TransitionDirection.cs" />
     <Compile Include="Enums\Tr4Level.cs" />
     <Compile Include="Enums\Tr4LevelSection.cs" />
