+++ Util/EnumExtensions.cs
@@ -2,7 +2,7 @@
 using System.ComponentModel;
 using System.Reflection;
 
-namespace TR4;
+namespace Util;
 
 public static class EnumExtensions
 {
